Feat S3 파일 업로드

+++ src/main/java/com/example/peernow360/dto/ReviewDto.java
@@ -20,5 +20,5 @@ public class ReviewDto {
     private int score5;
     private String reg_date;
     private int total;
-    private float avg;
+    private int avg;
 }
